HSEARCH-3398 Move object loading to SPI in a dedicated package
There was no point exposing ObjectLoader in APIs.
