+++ engine/src/main/java/org/hibernate/search/engine/search/loading/spi/IdentityObjectLoader.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.engine.search;
+package org.hibernate.search.engine.search.loading.spi;
 
 import java.util.List;
