+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/impl/FullTextQueryImpl.java
@@ -31,7 +31,7 @@ import org.hibernate.query.internal.ParameterMetadataImpl;
 import org.hibernate.query.spi.QueryParameterBindings;
 import org.hibernate.query.spi.ScrollableResultsImplementor;
 import org.hibernate.search.mapper.orm.hibernate.FullTextQuery;
-import org.hibernate.search.mapper.orm.search.impl.MutableObjectLoadingOptions;
+import org.hibernate.search.mapper.orm.search.loading.impl.MutableObjectLoadingOptions;
 import org.hibernate.search.engine.search.SearchQuery;
 import org.hibernate.transform.ResultTransformer;
 import org.hibernate.type.Type;
