+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/search/loading/impl/ByTypeObjectLoader.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.orm.search.impl;
+package org.hibernate.search.mapper.orm.search.loading.impl;
 
 import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
@@ -15,7 +15,7 @@ import java.util.Map;
 
 import org.hibernate.search.mapper.orm.logging.impl.Log;
 import org.hibernate.search.mapper.pojo.search.PojoReference;
-import org.hibernate.search.engine.search.ObjectLoader;
+import org.hibernate.search.engine.search.loading.spi.ObjectLoader;
 import org.hibernate.search.util.impl.common.LoggerFactory;
 
 public class ByTypeObjectLoader<O, T> implements ObjectLoader<PojoReference, T> {
