+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/search/loading/impl/ComposableObjectLoader.java
@@ -4,12 +4,12 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.orm.search.impl;
+package org.hibernate.search.mapper.orm.search.loading.impl;
 
 import java.util.List;
 import java.util.Map;
 
-import org.hibernate.search.engine.search.ObjectLoader;
+import org.hibernate.search.engine.search.loading.spi.ObjectLoader;
 
 /**
  * An {@link ObjectLoader} that can be easily composed with others object loaders.
